Convert Map into a special purpose SerialPortMap
Prefer delegation over inheritance, in order to be more specific about
what functions we want to expose. Move logic that is not needed in
the serialport interaction into serialportMap.
